Refactor logging.
